+++ lib/src/transitive_use/_exports.dart
@@ -1,0 +1,3 @@
+export 'package:dart_dependency_checker/src/transitive_use/transitive_use_checker.dart';
+export 'package:dart_dependency_checker/src/transitive_use/transitive_use_params.dart';
+export 'package:dart_dependency_checker/src/transitive_use/transitive_use_results.dart';