+++ test/resources/samples/all_sources_dirs_multi/lib/main.dart
@@ -1,2 +1,3 @@
 // ignore_for_file: depend_on_referenced_packages, unused_import
 import 'package:args/args.dart';
+import 'package:equatable/equatable.dart';