+++ test/resources/samples/meant_for_fixing/lib/main.dart
@@ -1,2 +1,2 @@
-// ignore_for_file: unused_import
+// ignore_for_file: depend_on_referenced_packages, unused_import
 import 'package:args/args.dart';