+++ test/src/transitive_use/transitive_use_results_test.dart
@@ -1,9 +1,9 @@
-import 'package:dart_dependency_checker/src/deps_unused/deps_unused_results.dart';
+import 'package:dart_dependency_checker/src/transitive_use/transitive_use_results.dart';
 import 'package:test/test.dart';
 
 void main() {
   test('has known props count', () {
-    const results = DepsUnusedResults(
+    const results = TransitiveUseResults(
       mainDependencies: {},
       devDependencies: {},
     );